fix: disconnect browser when config uses connect option
`disconnect` browser context in teardown phase instead of closing it when the setup phase uses `connect` options.  This allows for externally managed chrome services to be managed by seperate or alternative mechanism to support long-running or CI grids.

https://pptr.dev/#?product=Puppeteer&version=v1.11.0&show=api-browserdisconnect
```
Disconnects Puppeteer from the browser, but leaves the Chromium process running. After calling disconnect, the Browser object is considered disposed and cannot be used anymore.
```